pyproject.toml
[tool.poetry]
name = "eating_helper"
version = "0.1.0"
description = "Helps me remember to cook and eat well"
authors = ["Ben Yang <ben.yang@utexas.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
PyYAML = "^6.0"
google-api-python-client = "^2.66.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.1"
requests-cache = "^0.9.7"
tqdm = "^4.64.1"
gcsa = "^2.1.0"
pydantic = "^2.1.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
view = "eating_helper.main:view"
groc = "eating_helper.main:grocery"
cal= "eating_helper.main:calendar"
test_google_tasks_api = "eating_helper.google_api.tasks:get_google_tasks_service"
test_google_calendar_api = "eating_helper.google_api.calendar:test"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"
